el/requirements.txt
psycopg2-binary
google-cloud-bigquery
pandas
db-dtypes
